refactor OrderTests directories and file names
