+++ src/backend/OrderTests/Actions/CreateOrder/CreateOrderAddAddressTest.cs
@@ -3,20 +3,17 @@ using Core;
 using Core.Database;
 using Core.Database.Seeders;
 using Core.Exceptions;
-using Core.Setup.Auth0;
 using Core.Setup.Enums;
 using FluentValidation;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Orders;
 using Orders.Actions.CreateOrderActions.CreateOrderAddAddress;
-using OrdersTests.Actions.RequestHelpActionTest;
-using OrderTests.Actions.CreateOrderActionTest.CreateOrderDeleteAddressActionTest;
 using TestsBase;
 
-namespace OrdersTests.Actions.CreateOrderActionTest.CreateOrderAddAddressActionTest
+namespace OrderTests.Actions.CreateOrder
 {
     [TestClass()]
-    public class CreateOrderAddAddressActionTest
+    public class CreateOrderAddAddressTest
     {
         private readonly List<Module> _usedModules = new()
         {
@@ -25,7 +22,7 @@ namespace OrdersTests.Actions.CreateOrderActionTest.CreateOrderAddAddressActionT
         };
 
         [TestMethod()]
-        public async Task CreateOrderAddAddressActionTest_AddAddress_AddsAddressToDatabase()
+        public async Task CreateOrderAddAddressTest_AddAddress_AddsAddressToDatabase()
         {
             using var toolkit = new TestsToolkit(_usedModules);
             var context = toolkit.Resolve<AppDbContext>();
@@ -34,10 +31,10 @@ namespace OrdersTests.Actions.CreateOrderActionTest.CreateOrderAddAddressActionT
             // Arrange 
             var country = "Ukraine";
 
-            var testingUser1 = CreateOrderAddAddressActionFixture.CreateUser("Normal", "User");
-            var testingAuth0User1 = CreateOrderAddAddressActionFixture.CreateAuth0User(testingUser1, 1);
-            var testUserRole1 = CreateOrderAddAddressActionFixture.CreateUserRole(testingUser1, RoleSeeder.Role3Needy.Id);
-            var newPayload = CreateOrderAddAddressActionFixture.CreatePayload(country);
+            var testingUser1 = CreateOrderAddAddressFixture.CreateUser("Normal", "User");
+            var testingAuth0User1 = CreateOrderAddAddressFixture.CreateAuth0User(testingUser1, 1);
+            var testUserRole1 = CreateOrderAddAddressFixture.CreateUserRole(testingUser1, RoleSeeder.Role3Needy.Id);
+            var newPayload = CreateOrderAddAddressFixture.CreatePayload(country);
 
             context.Add(testingUser1);
             context.Add(testingAuth0User1);
@@ -45,7 +42,7 @@ namespace OrdersTests.Actions.CreateOrderActionTest.CreateOrderAddAddressActionT
 
             await context.SaveChangesAsync();
 
-            toolkit.UpdateUserInfo(CreateOrderAddAddressActionFixture.GetCurrentUserInfo(testingAuth0User1));
+            toolkit.UpdateUserInfo(CreateOrderAddAddressFixture.GetCurrentUserInfo(testingAuth0User1));
 
             // Act
             var executed = await action.Execute(newPayload);
@@ -58,7 +55,7 @@ namespace OrdersTests.Actions.CreateOrderActionTest.CreateOrderAddAddressActionT
         }
 
         [TestMethod()]
-        public async Task CreateOrderAddAddressActionTest_AddAddressPayloadNullStreetValues_AddsAddressToDatabase()
+        public async Task CreateOrderAddAddressTest_AddAddressPayloadNullStreetValues_AddsAddressToDatabase()
         {
             using var toolkit = new TestsToolkit(_usedModules);
             var context = toolkit.Resolve<AppDbContext>();
@@ -66,10 +63,10 @@ namespace OrdersTests.Actions.CreateOrderActionTest.CreateOrderAddAddressActionT
 
             // Arrange 
             var country = "Poland";
-            var testingUser1 = CreateOrderAddAddressActionFixture.CreateUser("Normal", "User");
-            var testingAuth0User1 = CreateOrderAddAddressActionFixture.CreateAuth0User(testingUser1, 1);
-            var testUserRole1 = CreateOrderAddAddressActionFixture.CreateUserRole(testingUser1, RoleSeeder.Role3Needy.Id);
-            var newPayload = CreateOrderAddAddressActionFixture.CreatePayloadNullStreetValues(country);
+            var testingUser1 = CreateOrderAddAddressFixture.CreateUser("Normal", "User");
+            var testingAuth0User1 = CreateOrderAddAddressFixture.CreateAuth0User(testingUser1, 1);
+            var testUserRole1 = CreateOrderAddAddressFixture.CreateUserRole(testingUser1, RoleSeeder.Role3Needy.Id);
+            var newPayload = CreateOrderAddAddressFixture.CreatePayloadNullStreetValues(country);
 
             context.Add(testingUser1);
             context.Add(testingAuth0User1);
@@ -77,7 +74,7 @@ namespace OrdersTests.Actions.CreateOrderActionTest.CreateOrderAddAddressActionT
 
             await context.SaveChangesAsync();
 
-            toolkit.UpdateUserInfo(CreateOrderAddAddressActionFixture.GetCurrentUserInfo(testingAuth0User1));
+            toolkit.UpdateUserInfo(CreateOrderAddAddressFixture.GetCurrentUserInfo(testingAuth0User1));
 
             // Act
             var executed = await action.Execute(newPayload);
@@ -90,7 +87,7 @@ namespace OrdersTests.Actions.CreateOrderActionTest.CreateOrderAddAddressActionT
         }
 
         [TestMethod()]
-        public async Task CreateOrderAddAddressActionTest_EmptyPayload_ThrowsException()
+        public async Task CreateOrderAddAddressTest_EmptyPayload_ThrowsException()
         {
             using var toolkit = new TestsToolkit(_usedModules);
             var context = toolkit.Resolve<AppDbContext>();
@@ -99,9 +96,9 @@ namespace OrdersTests.Actions.CreateOrderActionTest.CreateOrderAddAddressActionT
             // Arrange 
             var emptyPayload = new CreateOrderAddAddressPayload();
 
-            var testingUser1 = CreateOrderAddAddressActionFixture.CreateUser("Normal", "User");
-            var testingAuth0User1 = CreateOrderAddAddressActionFixture.CreateAuth0User(testingUser1, 1);
-            var testUserRole1 = CreateOrderAddAddressActionFixture.CreateUserRole(testingUser1, RoleSeeder.Role3Needy.Id);
+            var testingUser1 = CreateOrderAddAddressFixture.CreateUser("Normal", "User");
+            var testingAuth0User1 = CreateOrderAddAddressFixture.CreateAuth0User(testingUser1, 1);
+            var testUserRole1 = CreateOrderAddAddressFixture.CreateUserRole(testingUser1, RoleSeeder.Role3Needy.Id);
 
             context.Add(testingUser1);
             context.Add(testingAuth0User1);
@@ -109,7 +106,7 @@ namespace OrdersTests.Actions.CreateOrderActionTest.CreateOrderAddAddressActionT
 
             await context.SaveChangesAsync();
 
-            toolkit.UpdateUserInfo(CreateOrderAddAddressActionFixture.GetCurrentUserInfo(testingAuth0User1));
+            toolkit.UpdateUserInfo(CreateOrderAddAddressFixture.GetCurrentUserInfo(testingAuth0User1));
 
             // Act
             var exception = await Assert.ThrowsExceptionAsync<ValidationException>(() => action.Execute(emptyPayload));
@@ -121,18 +118,18 @@ namespace OrdersTests.Actions.CreateOrderActionTest.CreateOrderAddAddressActionT
 
 
         [TestMethod()]
-        public async Task CreateOrderAddAddressActionTest_NoCountry_ThrowsException()
+        public async Task CreateOrderAddAddressTest_NoCountry_ThrowsException()
         {
             using var toolkit = new TestsToolkit(_usedModules);
             var context = toolkit.Resolve<AppDbContext>();
             var action = toolkit.Resolve<CreateOrderAddAddressAction>();
 
             // Arrange 
-            var newPayload = CreateOrderAddAddressActionFixture.CreatePayload("HeheLand");
+            var newPayload = CreateOrderAddAddressFixture.CreatePayload("HeheLand");
 
-            var testingUser1 = CreateOrderAddAddressActionFixture.CreateUser("Normal", "User");
-            var testingAuth0User1 = CreateOrderAddAddressActionFixture.CreateAuth0User(testingUser1, 1);
-            var testUserRole1 = CreateOrderDeleteAddressActionFixture.CreateUserRole(testingUser1, RoleSeeder.Role3Needy.Id);
+            var testingUser1 = CreateOrderAddAddressFixture.CreateUser("Normal", "User");
+            var testingAuth0User1 = CreateOrderAddAddressFixture.CreateAuth0User(testingUser1, 1);
+            var testUserRole1 = CreateOrderDeleteAddressFixture.CreateUserRole(testingUser1, RoleSeeder.Role3Needy.Id);
 
             context.Add(testingUser1);
             context.Add(testingAuth0User1);
@@ -140,7 +137,7 @@ namespace OrdersTests.Actions.CreateOrderActionTest.CreateOrderAddAddressActionT
 
             await context.SaveChangesAsync();
 
-            toolkit.UpdateUserInfo(CreateOrderAddAddressActionFixture.GetCurrentUserInfo(testingAuth0User1));
+            toolkit.UpdateUserInfo(CreateOrderAddAddressFixture.GetCurrentUserInfo(testingAuth0User1));
 
             // Act
             var exception = await Assert.ThrowsExceptionAsync<ItemNotFoundException>(() => action.Execute(newPayload));
