+++ src/backend/OrderTests/Actions/CreateOrder/CreateOrderCreateOrderFixture.cs
@@ -1,25 +1,25 @@
 ﻿using Core.Database.Enums;
 using Core.Database.Models.Auth;
-using Core.Database.Models.Core;
 using Core.Database.Seeders;
 using Core.Setup.Auth0;
 
-namespace OrderTests.Actions.CreateOrderActionTest.CreateOrderGetAddressesActionTest
+namespace OrderTests.Actions.CreateOrder
 {
-    public static class CreateOrderGetAddressesActionFixture
+    public static class CreateOrderCreateOrderFixture
     {
-        public static User CreateUser(string firstName, string lastName) => new()
+        public static User CreateUser(string firstName, string lastName, long? warehouseId = null) => new()
         {
             Status = DbEntityStatus.Active,
             FirstName = firstName,
             LastName = lastName,
             Email = firstName + "@" + lastName + ".com",
+            WarehouseId = warehouseId
         };
 
         public static Auth0User CreateAuth0User(User user, int id) => new()
         {
-            FirstName = "Auth",
-            LastName = "Auth0User" + id,
+            FirstName = user.FirstName,
+            LastName = user.LastName + id,
             Nickname = "Auth0",
             UpdateUserId = UserSeeder.ServiceUser.Id,
             UpdatedAt = DateTime.UtcNow,
@@ -38,7 +38,6 @@ namespace OrderTests.Actions.CreateOrderActionTest.CreateOrderGetAddressesAction
             UpdatedAt = DateTime.UtcNow,
             Status = DbEntityStatus.Active
         };
-
         public static CurrentUserInfo GetCurrentUserInfo(Auth0User auth0User) => new()
         {
             Email = auth0User.Email,
@@ -50,31 +49,6 @@ namespace OrderTests.Actions.CreateOrderActionTest.CreateOrderGetAddressesAction
             Name = auth0User.FirstName + auth0User.LastName,
             Nickname = auth0User.Nickname,
             UpdatedAt = DateTime.UtcNow,
-        };
-
-        public static Address CreateAddress(Address addres1, Address address2) => new()
-        {
-            CountryId = 177,
-            PostalCode = addres1.PostalCode,
-            City = addres1.City,
-            Street = addres1.Street,
-            StreetNumber = address2.StreetNumber,
-            Apartment = address2.Apartment,
-            GpsLatitude = address2.GpsLatitude,
-            GpsLongitude = addres1.GpsLongitude,
-            UpdateUserId = UserSeeder.ServiceUser.Id,
-            UpdatedAt = DateTime.UtcNow,
-            Status = DbEntityStatus.Active
-        };
-
-        public static UserAddress CreateUserAddress(User user, Address address) => new()
-        {
-            User = user,
-            Address = address,
-            IsDeletedByUser = false,
-            UpdateUser = user,
-            UpdatedAt = DateTime.UtcNow,
-            Status = DbEntityStatus.Active
         };
     }
 }
