+++ src/backend/OrderTests/Actions/CreateOrder/CreateOrderDeleteAddressFixture.cs
@@ -1,11 +1,12 @@
 ﻿using Core.Database.Enums;
 using Core.Database.Models.Auth;
+using Core.Database.Models.Core;
 using Core.Database.Seeders;
 using Core.Setup.Auth0;
 
-namespace OrderTests.Actions.CreateOrderActionTest.CreateOrderGetCountriesActionTest
+namespace OrderTests.Actions.CreateOrder
 {
-    public static class CreateOrderGetCountriesFixture
+    public static class CreateOrderDeleteAddressFixture
     {
         public static User CreateUser(string firstName, string lastName) => new()
         {
@@ -49,6 +50,31 @@ namespace OrderTests.Actions.CreateOrderActionTest.CreateOrderGetCountriesAction
             Name = auth0User.FirstName + auth0User.LastName,
             Nickname = auth0User.Nickname,
             UpdatedAt = DateTime.UtcNow,
+        };
+
+        public static Address CreateAddress(Address addres1, Address address2) => new()
+        {
+            CountryId = 177,
+            PostalCode = addres1.PostalCode,
+            City = addres1.City,
+            Street = addres1.Street,
+            StreetNumber = address2.StreetNumber,
+            Apartment = address2.Apartment,
+            GpsLatitude = address2.GpsLatitude,
+            GpsLongitude = addres1.GpsLongitude,
+            UpdateUserId = UserSeeder.ServiceUser.Id,
+            UpdatedAt = DateTime.UtcNow,
+            Status = DbEntityStatus.Active
+        };
+
+        public static UserAddress CreateUserAddress(User user, Address address) => new()
+        {
+            User = user,
+            Address = address,
+            IsDeletedByUser = false,
+            UpdateUser = user,
+            UpdatedAt = DateTime.UtcNow,
+            Status = DbEntityStatus.Active
         };
     }
 }
