+++ src/backend/OrderTests/Actions/RequestHelp/RequestHelpGetMapFixture.cs
@@ -2,11 +2,10 @@
 using Core.Database.Models.Auth;
 using Core.Database.Models.Core;
 using Core.Database.Seeders;
-using Core.Setup.Auth0;
 
-namespace OrderTests.Actions.CreateOrderActionTest.CreateOrderDeleteAddressActionTest
+namespace OrderTests.Actions.RequestHelp
 {
-    public static class CreateOrderDeleteAddressActionFixture
+    public static class RequestHelpGetMapFixture
     {
         public static User CreateUser(string firstName, string lastName) => new()
         {
@@ -32,47 +31,31 @@ namespace OrderTests.Actions.CreateOrderActionTest.CreateOrderDeleteAddressActio
         public static UserRole CreateUserRole(User user, long roleId) => new()
         {
             RoleId = roleId,
-            UserId = user.Id,
             User = user,
             UpdateUserId = UserSeeder.ServiceUser.Id,
             UpdatedAt = DateTime.UtcNow,
             Status = DbEntityStatus.Active
         };
 
-        public static CurrentUserInfo GetCurrentUserInfo(Auth0User auth0User) => new()
+        public static Order CreateOrder(User user, Address address, string orderName = "ORD000001") => new()
         {
-            Email = auth0User.Email,
-            EmailVerified = true,
-            Identifier = auth0User.Identifier,
-            GivenName = auth0User.FirstName,
-            FamilyName = auth0User.LastName,
-            Locale = "pl",
-            Name = auth0User.FirstName + auth0User.LastName,
-            Nickname = auth0User.Nickname,
-            UpdatedAt = DateTime.UtcNow,
-        };
-
-        public static Address CreateAddress(Address addres1, Address address2) => new()
-        {
-            CountryId = 177,
-            PostalCode = addres1.PostalCode,
-            City = addres1.City,
-            Street = addres1.Street,
-            StreetNumber = address2.StreetNumber,
-            Apartment = address2.Apartment,
-            GpsLatitude = address2.GpsLatitude,
-            GpsLongitude = addres1.GpsLongitude,
+            Name = orderName,
+            Percentage = 50,
+            AddressId = address.Id,
+            OwnerUser = user,
+            CreationDate = DateTime.UtcNow,
+            IsCanceledByUser = false,
             UpdateUserId = UserSeeder.ServiceUser.Id,
             UpdatedAt = DateTime.UtcNow,
             Status = DbEntityStatus.Active
         };
 
-        public static UserAddress CreateUserAddress(User user, Address address) => new()
+        public static OrderProduct CreateOrderProduct(Order order, long productId, int quantity) => new()
         {
-            User = user,
-            Address = address,
-            IsDeletedByUser = false,
-            UpdateUser = user,
+            Order = order,
+            ProductId = productId,
+            Quantity = quantity,
+            UpdateUserId = UserSeeder.ServiceUser.Id,
             UpdatedAt = DateTime.UtcNow,
             Status = DbEntityStatus.Active
         };
